Update stocks.py
wasnt getting an error yesterday bc the median was calculated correctly (not divis by 2)

+++ HW 2.5/stocks.py
@@ -92,7 +92,7 @@ def process_data(data:dict)->dict:
         mean = sum(prices)/length
         
         if length % 2:
-            median = prices[length / 2]
+            median = prices[int(length / 2)]
         else:
             median = prices[int(length / 2)] + prices[int((length / 2) - 1)] /2
